% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_de.R
\name{de_naive}
\alias{de_naive}
\title{Naive: subtract ambient point estimate, then test condition per gene.}
\usage{
de_naive(sim)
}
\arguments{
\item{sim}{output of \code{\link[=simulate_de]{simulate_de()}}.}
}
\value{
per-gene p-values.
}
\description{
Naive: subtract ambient point estimate, then test condition per gene.
}
