% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decant-package.R
\docType{package}
\name{Decant-package}
\alias{Decant}
\alias{Decant-package}
\title{Decant: benchmark-first ambient RNA decontamination}
\description{
Start with \code{\link[=decant]{decant()}} for correction, \code{\link[=split_droplets]{split_droplets()}} /
\code{\link[=read_10x_counts]{read_10x_counts()}} to get real data in, and the \verb{gate_*()} functions to
re-derive every default from ground truth.
}
\author{
\strong{Maintainer}: Kushal Raj Roy \email{kushalrajroy1@gmail.com} (\href{https://orcid.org/0000-0002-4664-7462}{ORCID})

}
\keyword{internal}
