% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decant.R
\name{print.decant}
\alias{print.decant}
\title{Print a decant result.}
\usage{
\method{print}{decant}(x, ...)
}
\arguments{
\item{x}{a \code{decant} object.}

\item{...}{unused.}
}
\description{
Print a decant result.
}
