% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{summarise_benchmark}
\alias{summarise_benchmark}
\title{Aggregate to mean +/- sd per (method, soup_bias).}
\usage{
summarise_benchmark(df)
}
\arguments{
\item{df}{output of \code{\link[=run_benchmark]{run_benchmark()}}.}
}
\description{
Aggregate to mean +/- sd per (method, soup_bias).
}
